+++ docs/changelog.md
@@ -2,6 +2,15 @@
 sidebarDepth: 0
 ---
 
+## 4.0.0
+* Add support for arbitrary resources, like Ingress or PodDisruptionBudget
+* Add support check templating (see docs)
+* Add support for multi-resource checks (see docs)
+
+### Breaking Changes
+* In custom checks, `jsonSchema` is now `schemaString`
+* Check `pdbDisruptionsAllowedGreaterThanZero` is now called `pdbDisruptionsIsZero`
+
 ## 3.2.0
 * Add `--format=pretty` option for CLI output